# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  github_com_EvolutionAPI_evolution-go_pkg_newsletter_service.CreateNewsletterStruct:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_newsletter_service.GetNewsletterInviteStruct:
    properties:
      key:
        type: string
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_newsletter_service.GetNewsletterMessagesStruct:
    properties:
      before_id:
        type: integer
      count:
        type: integer
      jid:
        $ref: '#/definitions/types.JID'
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_newsletter_service.GetNewsletterStruct:
    properties:
      jid:
        $ref: '#/definitions/types.JID'
    type: object
...
  types.JID:
    properties:
      device:
        format: int32
        type: integer
      integrator:
        format: int32
        type: integer
      rawAgent:
        format: int32
        type: integer
      server:
        type: string
      user:
        type: string
    type: object
...
  types.WebAuthnResponse:
    properties:
      id:
        type: string
      rawId:
        items:
          type: integer
        type: array
      response:
        $ref: '#/definitions/types.WebAuthnResponseData'
      type:
        type: string
    type: object
  types.WebAuthnResponseData:
    properties:
      authenticatorData:
        items:
          type: integer
        type: array
      clientDataJSON:
        items:
          type: integer
        type: array
      signature:
        items:
          type: integer
        type: array
      userHandle:
        items:
          type: integer
        type: array
    type: object
...
info:
  contact: {}
  description: Evolution GO - whatsmeow
  title: Evolution GO
  version: "1.0"
paths:
...
  /newsletter/create:
    post:
      consumes:
      - application/json
      description: Create newsletter
      parameters:
      - description: Newsletter data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_newsletter_service.CreateNewsletterStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create newsletter
      tags:
      - Newsletter
  /newsletter/info:
    post:
      consumes:
      - application/json
      description: Get newsletter
      parameters:
      - description: Newsletter data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_newsletter_service.GetNewsletterStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get newsletter
      tags:
      - Newsletter
  /newsletter/link:
    post:
      consumes:
      - application/json
      description: Get newsletter invite
      parameters:
      - description: Newsletter data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_newsletter_service.GetNewsletterInviteStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get newsletter invite
      tags:
      - Newsletter
  /newsletter/list:
    get:
      consumes:
      - application/json
      description: List newsletters
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List newsletters
      tags:
      - Newsletter
  /newsletter/messages:
    post:
      consumes:
      - application/json
      description: Get newsletter messages
      parameters:
      - description: Newsletter data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_newsletter_service.GetNewsletterMessagesStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get newsletter messages
      tags:
      - Newsletter
  /newsletter/subscribe:
    post:
      consumes:
      - application/json
      description: Subscribe newsletter
      parameters:
      - description: Newsletter data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_newsletter_service.GetNewsletterStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Subscribe newsletter
      tags:
      - Newsletter
  /passkey-ceremony/{token}:
    get:
      description: Returns the current WebAuthn passkey-pairing ceremony state for
        a token. PUBLIC endpoint (no apikey) — access is gated by the opaque short-lived
        ceremony token. Polled by the Evolution Passkey Helper browser extension.
      parameters:
      - description: Ceremony token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ceremony state ({stage, skipHandoffUX, publicKey?, code?, error?})
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: token is required
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: ceremony not found or expired
          schema:
            $ref: '#/definitions/gin.H'
        "503":
          description: passkey ceremony unavailable
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get passkey ceremony state
      tags:
      - Passkey
  /passkey-ceremony/{token}/confirm:
    post:
      description: Finishes the passkey pairing after the user verified the confirmation
        code. PUBLIC endpoint (no apikey) — gated by the ceremony token.
      parameters:
      - description: Ceremony token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: token is required
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: ceremony not found or expired
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
        "503":
          description: passkey ceremony unavailable
          schema:
            $ref: '#/definitions/gin.H'
      summary: Confirm passkey pairing
      tags:
      - Passkey
  /passkey-ceremony/{token}/response:
    post:
      consumes:
      - application/json
      description: Receives the WebAuthn assertion produced by the browser extension
        and forwards it to WhatsApp. PUBLIC endpoint (no apikey) — gated by the ceremony
        token. Body is the WebAuthnResponse shape (id, rawId, type, response{clientDataJSON,
        authenticatorData, signature, userHandle?}), base64url-unpadded.
      parameters:
      - description: Ceremony token
        in: path
        name: token
        required: true
        type: string
      - description: WebAuthn assertion
        in: body
        name: response
        required: true
        schema:
          $ref: '#/definitions/types.WebAuthnResponse'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: token is required / invalid body
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: ceremony not found or expired
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
        "503":
          description: passkey ceremony unavailable
          schema:
            $ref: '#/definitions/gin.H'
      summary: Submit passkey WebAuthn response
      tags:
      - Passkey
...
swagger: "2.0"